Extract layergroup-metadata class
